fix: move AbstractMetadataService up one dir
